% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_targets.r
\name{target_prob_hivdx_byraceage}
\alias{target_prob_hivdx_byraceage}
\alias{target_prob_agedx_byrace}
\alias{target_prob_vls_byraceage}
\alias{target_prob_prep_byrace}
\alias{target_prop_anatsites_tested}
\alias{target_prob_gcpos_tested_anatsites}
\title{Calculating Target Statistics}
\usage{
target_prob_hivdx_byraceage(dat, tailn)

target_prob_agedx_byrace(dat, tailn)

target_prob_vls_byraceage(dat, tailn)

target_prob_prep_byrace(dat, tailn)

target_prop_anatsites_tested(dat, tailn)

target_prob_gcpos_tested_anatsites(dat, tailn)
}
\arguments{
\item{dat}{An object of class netsim.}

\item{tailn}{The number of weeks (always at the tail) to extract and average over.}
}
\description{
These functions extract and calculate simulated outcomes
based on the calibration target statistics specified.
}
\section{Functions}{
\itemize{
\item \code{target_prob_hivdx_byraceage}: Probability of being diagnosed with HIV among those with HIV, by race and age.

\item \code{target_prob_agedx_byrace}: Calculate age distribution among
those diagnosed with HIV, by race/ethnicity.

\item \code{target_prob_vls_byraceage}: Calculate probability of viral suppression among those with diagnosed HIV, by race/ethnicity and age.

\item \code{target_prob_prep_byrace}: Calculate probability of PrEP use among indicated agents, by race/ethnicity.

\item \code{target_prop_anatsites_tested}: Return proportion of anatomic sites tested among MSM seeking testing at the clinic.

\item \code{target_prob_gcpos_tested_anatsites}: Return probability of being GC+ at tested anatomic sites.
}}

