% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{raceage_grid}
\alias{raceage_grid}
\alias{test_xgc}
\alias{setenv}
\alias{get_est}
\alias{param_sets}
\title{Helper functions}
\usage{
raceage_grid()

test_xgc(seed = 123, priorlist = priors)

setenv(
  nsteps = 200,
  tail_length = 52,
  sims_below_tol = 50,
  slurm_nprocs = Sys.getenv("SLURM_NPROCS")
)

get_est(string = c("epistats", "netest", "netstats"))

param_sets(num_sets, priorlist)
}
\arguments{
\item{priorlist}{A prior list generated for use in the EasyABC sequential algorithm.}

\item{num_sets}{Number of parameter sets to generate.}
}
\description{
These functions provide easy access to demographic distributions
in the target population as well as helper functions to generate lookup
tables for different combinations of parameter levels.
}
\details{
'get_est' checks for existence of ARTnet code repo. If so,
it checks which project's copy of a file is newer. If the version
in ARTnet code repo is newer, save it to this project and load. Otherwise,
load the existing local file. The function is intended to ensure we're always
working with the most recent file. Will load the local file by default
if the ARTnet repo's file directories are not found (so that scripts contained
in this package won't fail in the absence of the ARTnet code repo.)
}
\section{Functions}{
\itemize{
\item \code{raceage_grid}: Produce a lookup table of all possible combinations of
race/ethnicity and age group.

\item \code{test_xgc}: Helper function to use as input into the xgc() function

\item \code{setenv}: Helper function to set temporary environment variable.

\item \code{get_est}: Retrieve epistats object.

\item \code{param_sets}: Make parameter sets
}}

