% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{raceage_grid}
\alias{raceage_grid}
\alias{test_xgc}
\alias{setenv}
\title{Helper functions}
\usage{
raceage_grid()

test_xgc(seed = 123, priorlist = priors)

setenv(
  nsteps = 200,
  tail_length = 52,
  sims_below_tol = 50,
  slurm_nprocs = Sys.getenv("SLURM_NPROCS")
)
}
\description{
These functions provide easy access to demographic distributions
in the target population as well as helper functions to generate lookup
tables for different combinations of parameter levels.
}
\section{Functions}{
\itemize{
\item \code{raceage_grid}: Produce a lookup table of all possible combinations of
race/ethnicity and age group.

\item \code{test_xgc}: Helper function to use as input into the xgc() function

\item \code{setenv}: Helper function to set temporary environment variable.
}}

